refactor(merge): Fixing additional merge conflicts.
